Update 4 files\n\nCo-authored-by: Freebuff Agent <agent@mail.freebuff.app>

+++ app/src/main/java/com/kennyb1201/kbstream/ui/player/NativePlayerActivity.kt
@@ -4818,7 +4818,7 @@ class NativePlayerActivity : ComponentActivity() {
         btnNextPlay.requestFocus()
 
         if (autoPlayNext) {
-            val threshold = AppPreferences.getStillThereEpisodes(this)
+            val threshold = AppPreferences.getStillThereEpisodes(this).toInt()
             val autoAdvanced = AppPreferences.getConsecutiveAutoplays(this)
             if (AppPreferences.getStillTherePrompt(this) && autoAdvanced >= threshold) {
                 // Binge watchdog: enough unattended episodes have played in a
